#117 add podspec dependency version constraints (#118)
Without version constraints, the underlying ios dependencies can update to
unsupported versions at any time. On 8/11/21 they did just that.
TextRecognition was bumped to 1.4.0 which happened to change the
MLKitTextRecognition.h header so that it no longer includes several necessary
imports. Version constraints should be explicitly made in this library's
podspec so that functionality can be guaranteed.